React 19 improvements
Split mirror pages into server/client components,
fetch data server-side and resolve via React.use()

Add Suspense fallbacks to layouts, error boundary for mirror detail

Replace setLoading callbacks with useTransition in form handlers

Use <Context> directly instead of <Context.Provider>
