+++ ui/app/mirrors/[mirrorId]/error.tsx
@@ -1,0 +1,6 @@
+'use client';
+import NoMirror from './nomirror';
+
+export default function MirrorError() {
+  return <NoMirror />;
+}
