+++ ui/app/peers/layout.tsx
@@ -2,7 +2,8 @@
 import { Configuration } from '@/app/config/config';
 import SidebarComponent from '@/components/SidebarComponent';
 import { Layout } from '@/lib/Layout';
-import { PropsWithChildren } from 'react';
+import { ProgressCircle } from '@/lib/ProgressCircle';
+import { PropsWithChildren, Suspense } from 'react';
 
 export default async function PageLayout({ children }: PropsWithChildren) {
   return (
@@ -13,7 +14,11 @@ export default async function PageLayout({ children }: PropsWithChildren) {
         />
       }
     >
-      {children}
+      <Suspense
+        fallback={<ProgressCircle variant='determinate_progress_circle' />}
+      >
+        {children}
+      </Suspense>
     </Layout>
   );
 }
