+++ ui/app/scripts/layout.tsx
@@ -2,6 +2,7 @@
 import { Configuration } from '@/app/config/config';
 import SidebarComponent from '@/components/SidebarComponent';
 import { Layout } from '@/lib/Layout';
+import { ProgressCircle } from '@/lib/ProgressCircle';
 import { PropsWithChildren, Suspense } from 'react';
 
 export default async function PageLayout({ children }: PropsWithChildren) {
@@ -13,7 +14,11 @@ export default async function PageLayout({ children }: PropsWithChildren) {
         />
       }
     >
-      <Suspense>{children}</Suspense>
+      <Suspense
+        fallback={<ProgressCircle variant='determinate_progress_circle' />}
+      >
+        {children}
+      </Suspense>
     </Layout>
   );
 }
